# docs/swagger.yaml
basePath: /api
definitions:
...
  github_com_jenish-brainztechs_go-backend_internal_adapter_handler_http_dto.CreateReportRequest:
    properties:
      report_no:
        type: string
      visit_id:
        type: string
    required:
    - report_no
    type: object
...
  github_com_jenish-brainztechs_go-backend_internal_adapter_handler_http_dto.CreateUser:
    properties:
      email:
        description: the acceptable string type is of email type
        type: string
      first_name:
        type: string
      last_name:
        type: string
      password:
        type: string
      phone:
        type: string
      role_name:
        type: string
    required:
    - email
    - first_name
    - last_name
    - password
    - role_name
    type: object
...
  github_com_jenish-brainztechs_go-backend_internal_adapter_handler_http_dto.ReportPrintRequest:
    properties:
      copy_number:
        type: integer
    required:
    - copy_number
    type: object
  github_com_jenish-brainztechs_go-backend_internal_adapter_handler_http_dto.ReportPrintResponse:
    properties:
      copy_number:
        type: integer
      id:
        type: string
      printed_at:
        type: string
      printed_by:
        type: string
      report_id:
        type: string
    type: object
  github_com_jenish-brainztechs_go-backend_internal_adapter_handler_http_dto.ReportResponse:
    properties:
      generated_at:
        type: string
      generated_by:
        type: string
      id:
        type: string
      pdf_path:
        type: string
      report_no:
        type: string
      status:
        $ref: '#/definitions/github_com_jenish-brainztechs_go-backend_internal_core_domain.ReportStatus'
      visit_id:
        type: string
    type: object
...
  github_com_jenish-brainztechs_go-backend_internal_adapter_handler_http_dto.UpdateReportRequest:
    properties:
      pdf_path:
        type: string
      status:
        $ref: '#/definitions/github_com_jenish-brainztechs_go-backend_internal_core_domain.ReportStatus'
    type: object
...
  github_com_jenish-brainztechs_go-backend_internal_core_domain.ReportStatus:
    enum:
    - draft
    - final
    - amended
    type: string
    x-enum-varnames:
    - ReportDraft
    - ReportFinal
    - ReportAmended
...
  internal_adapter_handler_http.errorResponse:
    properties:
      messages:
        example:
        - Error message 1
        - ' Error message 2'
        items:
          type: string
        type: array
      success:
        example: false
        type: boolean
    type: object
  internal_adapter_handler_http.response:
    properties:
      data: {}
      message:
        example: Success
        type: string
      success:
        example: true
        type: boolean
    type: object
host: localhost:8080
info:
  contact:
    email: support@example.com
    name: API Support
  description: Laboratory Information System API
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
  termsOfService: http://swagger.io/terms/
  title: LIS Backend API
  version: "1.0"
paths:
...
  /report:
    post:
      consumes:
      - application/json
      description: Create a new report for a visit
      parameters:
      - description: Report details
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/github_com_jenish-brainztechs_go-backend_internal_adapter_handler_http_dto.CreateReportRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/internal_adapter_handler_http.response'
            - properties:
                data:
                  $ref: '#/definitions/github_com_jenish-brainztechs_go-backend_internal_adapter_handler_http_dto.ReportResponse'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/internal_adapter_handler_http.errorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/internal_adapter_handler_http.errorResponse'
      security:
      - BearerAuth: []
      summary: Create report
      tags:
      - Reports
  /report/{id}:
    get:
      consumes:
      - application/json
      description: Get a report's details by ID
      parameters:
      - description: Report ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/internal_adapter_handler_http.response'
            - properties:
                data:
                  $ref: '#/definitions/github_com_jenish-brainztechs_go-backend_internal_adapter_handler_http_dto.ReportResponse'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/internal_adapter_handler_http.errorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/internal_adapter_handler_http.errorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/internal_adapter_handler_http.errorResponse'
      security:
      - BearerAuth: []
      summary: Get report by ID
      tags:
      - Reports
    patch:
      consumes:
      - application/json
      description: Update an existing report
      parameters:
      - description: Report ID
        in: path
        name: id
        required: true
        type: string
      - description: Report details
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/github_com_jenish-brainztechs_go-backend_internal_adapter_handler_http_dto.UpdateReportRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/internal_adapter_handler_http.response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/internal_adapter_handler_http.errorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/internal_adapter_handler_http.errorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/internal_adapter_handler_http.errorResponse'
      security:
      - BearerAuth: []
      summary: Update report
      tags:
      - Reports
  /report/{id}/print:
    post:
      consumes:
      - application/json
      description: Create a print record for a report
      parameters:
      - description: Report ID
        in: path
        name: id
        required: true
        type: string
      - description: Print details
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/github_com_jenish-brainztechs_go-backend_internal_adapter_handler_http_dto.ReportPrintRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/internal_adapter_handler_http.response'
            - properties:
                data:
                  $ref: '#/definitions/github_com_jenish-brainztechs_go-backend_internal_adapter_handler_http_dto.ReportPrintResponse'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/internal_adapter_handler_http.errorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/internal_adapter_handler_http.errorResponse'
      security:
      - BearerAuth: []
      summary: Create report print
      tags:
      - Reports
  /report/{id}/prints:
    get:
      consumes:
      - application/json
      description: Get all print records for a given report
      parameters:
      - description: Report ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/internal_adapter_handler_http.response'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/github_com_jenish-brainztechs_go-backend_internal_adapter_handler_http_dto.ReportPrintResponse'
                  type: array
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/internal_adapter_handler_http.errorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/internal_adapter_handler_http.errorResponse'
      security:
      - BearerAuth: []
      summary: List report prints
      tags:
      - Reports
  /report/visit/{visit_id}:
    get:
      consumes:
      - application/json
      description: Get all reports for a given visit
      parameters:
      - description: Visit ID
        in: path
        name: visit_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/internal_adapter_handler_http.response'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/github_com_jenish-brainztechs_go-backend_internal_adapter_handler_http_dto.ReportResponse'
                  type: array
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/internal_adapter_handler_http.errorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/internal_adapter_handler_http.errorResponse'
      security:
      - BearerAuth: []
      summary: List reports by visit
      tags:
      - Reports
...
  /user/create:
    post:
      consumes:
      - application/json
      description: Create a new user with role and profile
      parameters:
      - description: User details
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/github_com_jenish-brainztechs_go-backend_internal_adapter_handler_http_dto.CreateUser'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/internal_adapter_handler_http.response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/internal_adapter_handler_http.errorResponse'
        "409":
          description: Conflict
          schema:
            $ref: '#/definitions/internal_adapter_handler_http.errorResponse'
      summary: Create user
      tags:
      - Users
...
securityDefinitions:
  BearerAuth:
    description: 'Enter the token with the `Bearer: ` prefix, e.g. "Bearer abcde12345".'
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"